// src/frontend/src/api/model/currentUserResponseDto.ts
/**
 * ExampleWebApp API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CurrentUserStatus } from './currentUserStatus';
import { UserPermission } from './userPermission';


/**
 * M:ExampleWebApp.Backend.WebApi.AuthController.CurrentUser api response data.
 */
export interface CurrentUserResponseDto { 
    status: CurrentUserStatus;
    /**
     * Login username.
     */
    userName?: string | null;
    /**
     * Email address.
     */
    email?: string | null;
    /**
     * List of roles associated to this user.
     */
    roles?: Set<string> | null;
    /**
     * Permissions related to this user roles.
     */
    permissions?: Set<UserPermission> | null;
}
export namespace CurrentUserResponseDto {
}